+++ pybamm/models/submodels/active_material/__init__.py
@@ -1,4 +1,4 @@
 from .base_active_material import BaseModel
 from .constant_active_material import Constant
-from .varying_active_material_full import VaryingFull
-from .varying_active_material_uniform import VaryingUniform
+from .stress_driven_active_material import StressDriven
+from .reaction_driven_active_material import ReactionDriven