+++ tests/unit/test_citations.py
@@ -165,7 +165,12 @@ class TestCitations(unittest.TestCase):
 
         citations._reset()
         self.assertNotIn("Reniers2019", citations._papers_to_cite)
-        pybamm.active_material.VaryingFull(None, None, None)
+        pybamm.active_material.StressDriven(None, None, None, True)
+        self.assertIn("Reniers2019", citations._papers_to_cite)
+
+        citations._reset()
+        self.assertNotIn("Reniers2019", citations._papers_to_cite)
+        pybamm.active_material.ReactionDriven(None, None, None, True)
         self.assertIn("Reniers2019", citations._papers_to_cite)
 
     def test_parameter_citations(self):